Raisr_Opencl: Fix green edge problem
When we map vaSurface to cl_mem, media-driver may pad the image because
of the hardware alignment requirement. When we sample the pixel at
border Opencl may read data at padded area. Because We don't know if the
padded area is initialized or not, we'd better manually set the border
pixel value.

Signed-off-by: Wenbin Chen <wenbin.chen@intel.com>